+++ cmd/main.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"os"
 	"os/signal"
-	"sync"
 	"syscall"
 
 	"github.com/voltgizerz/POS-restaurant/config"
@@ -12,15 +11,13 @@ import (
 	"github.com/voltgizerz/POS-restaurant/internal/app/api/handler"
 	"github.com/voltgizerz/POS-restaurant/internal/app/api/middleware"
 	"github.com/voltgizerz/POS-restaurant/internal/app/auth"
-	"github.com/voltgizerz/POS-restaurant/internal/app/interactor"
 	"github.com/voltgizerz/POS-restaurant/internal/app/repository"
-	"github.com/voltgizerz/POS-restaurant/internal/app/service"
+	"github.com/voltgizerz/POS-restaurant/internal/core/interactor"
+	"github.com/voltgizerz/POS-restaurant/internal/core/service"
 	"github.com/voltgizerz/POS-restaurant/internal/database"
 	"github.com/voltgizerz/POS-restaurant/pkg/jeager"
 	"github.com/voltgizerz/POS-restaurant/pkg/logger"
 )
-
-var wg sync.WaitGroup
 
 func main() {
 	logger.Init()
