+++ internal/app/api/handler/auth_handler.go
@@ -8,10 +8,10 @@ import (
 	"github.com/opentracing/opentracing-go"
 
 	"github.com/voltgizerz/POS-restaurant/internal/app/api/common"
-	"github.com/voltgizerz/POS-restaurant/internal/app/entity"
-	"github.com/voltgizerz/POS-restaurant/internal/app/interactor"
-	"github.com/voltgizerz/POS-restaurant/internal/app/ports"
 	"github.com/voltgizerz/POS-restaurant/internal/constants"
+	"github.com/voltgizerz/POS-restaurant/internal/core/entity"
+	"github.com/voltgizerz/POS-restaurant/internal/core/interactor"
+	"github.com/voltgizerz/POS-restaurant/internal/core/ports"
 	"github.com/voltgizerz/POS-restaurant/internal/utils"
 )
