+++ internal/app/api/handler/auth_handler_test.go
@@ -9,9 +9,9 @@ import (
 	"github.com/bytedance/sonic"
 	"github.com/gofiber/fiber/v3"
 	"github.com/valyala/fasthttp"
-	"github.com/voltgizerz/POS-restaurant/internal/app/entity"
-	"github.com/voltgizerz/POS-restaurant/internal/app/interactor"
+	"github.com/voltgizerz/POS-restaurant/internal/core/interactor"
 	"github.com/voltgizerz/POS-restaurant/internal/constants"
+	"github.com/voltgizerz/POS-restaurant/internal/core/entity"
 	"go.uber.org/mock/gomock"
 )
