+++ internal/app/api/server_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/voltgizerz/POS-restaurant/internal/app/api/middleware"
-	"github.com/voltgizerz/POS-restaurant/internal/app/interactor"
+	"github.com/voltgizerz/POS-restaurant/internal/core/interactor"
 )
 
 func TestNewServer(t *testing.T) {
