+++ internal/app/repository/user_repository.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/jmoiron/sqlx"
 	"github.com/opentracing/opentracing-go"
-	"github.com/voltgizerz/POS-restaurant/internal/app/entity"
-	"github.com/voltgizerz/POS-restaurant/internal/app/ports"
+	"github.com/voltgizerz/POS-restaurant/internal/core/ports"
+	"github.com/voltgizerz/POS-restaurant/internal/core/entity"
 )
 
 const (
