+++ internal/core/entity/auth_entity.go
