+++ internal/core/entity/menu_entity.go
