+++ internal/core/entity/user_entity.go
