+++ internal/core/ports/menu_ports.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 
 	"github.com/gofiber/fiber/v3"
-	"github.com/voltgizerz/POS-restaurant/internal/app/entity"
+	"github.com/voltgizerz/POS-restaurant/internal/core/entity"
 )
 
 //go:generate mockgen -source=./internal/app/ports/menu_ports.go -destination=./internal/mocks/mocks_menu.go -package=mocks
