+++ internal/mocks/mocks_user.go
@@ -13,7 +13,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	entity "github.com/voltgizerz/POS-restaurant/internal/app/entity"
+	entity "github.com/voltgizerz/POS-restaurant/internal/core/entity"
 	gomock "go.uber.org/mock/gomock"
 )
